refactor: structure component
