+++ src/components/offer/offer-card/const.ts
