+++ src/components/offer/rating-view/rating-view.tsx
