+++ src/components/offer/review-view/review-view.tsx
@@ -1,5 +1,5 @@
-import { Review } from '../../mocks/types/reviews';
-import { formatDate } from '../../utils';
+import { Review } from '../../../mocks/types/reviews';
+import { formatDate } from '../../../utils';
 import RatingView from '../rating-view/rating-view';
 
 type ReviewViewProps = {
